Fix RecordNotFound exception in Users#{learnmore,landing_page}
Model#find will raise an exception if the record can't be found, so the
`thing || 'default text'` won't work -- it's either `thing` or an
exception is raised. This commit changes it to use `try` so that if the
record can't be found, it will fall back to the default text.